+++ hutool-core/src/test/java/cn/hutool/core/math/CalculatorTest.java
@@ -48,4 +48,11 @@ public class CalculatorTest {
 		final double conversion = Calculator.conversion("((-2395+0) * 0.3+140.24+35+90)/30");
 		Assert.assertEquals(-15.11, conversion, 0.01);
 	}
+
+	@Test
+	public void issue2964Test() {
+		// https://github.com/dromara/hutool/issues/2964
+		final double calcValue = Calculator.conversion("(11+2)12");
+		Assert.assertEquals(156D, calcValue, 0.001);
+	}
 }